Added Catalog writing, refactored all write tests
EasyGlue now supports writing to catalog via the table() and catalog() methods of EasyDynamicFrameWriter. This commit also includes several changes to the structure of writer tests, mainly now all tests inherit from the EasyGlueTest and TestWriter classes so that test fixtures are properly organized and instantiated. In the future reader tests will have to do the same.
 - Added CatalogMixin to the writer module
 - Added table() and catalog() methods as part of CatalogMixin
 - Added tests for Catalog writing under test_catalog
 - Moved qualified name validation function from reader.catalog package to utils package, so that writer methods can also use it
 - Renamed test.utils to test.test_utils
 - Created new test.utils package that will test easyglue.utils
 - Now all writer tests are subclasses of TestWriter, which is a subclass of EasyGlueTest. This improves fixture structure and conforms to unittest standards